Long Abscence Although Functioning
Everything is working to 95%, which i beleive is good enough for most
people. Only improvements to be made are running under smaller file
size, as it takes up around 80% of dynamic memory.

NOSERIAL file does not work at the moment, in attempt to lighten the
load on the arduino i removed the computer serial connection but it does
not write to SD if there is no serial. Im probably making a juvenile
mistake in my code.